Combine purchaser_name together with purchaser_lines
Quite often the purchaser name looks like:

~~~
John Doe,
Foo Inc.
~~~

Where the invoice is prepared for John Doe from Foo Inc. company. If
this is the case, the `purchaser_name` becomes too long and ovelaps the
`purchaser_lines.

The overlapping can be avoided by combining the `purchaser_name` and
`purchaser_lines`, outputting just the first line in bold and treating
the rest just as an address. This should provide the greatest
flexibility and it is actually extension to #77.