Replace breakdown subtable by line spanning the whole table
The nested tables do not work satisfactory.

Mainly, there seem to be issues with properly computing the headers
widht, which sometime are wrapped, while there is sill plenty of space.

Secondly, the lenght of break down line is quite limited, while it does
not really harm if it exceeds to the nehboring colum. Even if that was
problem, the entered text can be wrapped to multiple lines.

Given these limitations, it is better to emulate single entries by
multiple lines, where the firs line, which contains the important
information is split into the columns, while the breakdown spans the
whole table.